Removed old OrchestrationManager
