+++ Fightship Arena/Assets/Scripts/Managers/Levels/LevelManager.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 using FightShipArena.Assets.Scripts.Managers.EnemyManagement;
 using FightShipArena.Assets.Scripts.Managers.HudManagement;
 using FightShipArena.Assets.Scripts.Managers.OrchestrationManagement;
-using FightShipArena.Assets.Scripts.Managers.OrchestrationManagement2;
 using FightShipArena.Assets.Scripts.Managers.SceneManagement;
 using FightShipArena.Assets.Scripts.Managers.ScoreManagement;
 using FightShipArena.Assets.Scripts.Managers.SoundManagement;
