+++ Fightship Arena/Assets/Scripts/Managers/Levels/Level_01Manager.cs
@@ -5,7 +5,6 @@ using System.Text;
 using System.Threading.Tasks;
 using FightShipArena.Assets.Scripts.Managers.EnemyManagement;
 using FightShipArena.Assets.Scripts.Managers.OrchestrationManagement;
-using FightShipArena.Assets.Scripts.Managers.OrchestrationManagement2;
 using FightShipArena.Assets.Scripts.Managers.ScoreManagement;
 using FightShipArena.Assets.Scripts.Managers.SoundManagement;
 using FightShipArena.Assets.Scripts.Player;
