+++ Fightship Arena/Assets/Scripts/Managers/OrchestrationManagement/OrchestrationManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0deadee27eaae7147812a46a10a6bdc5
+guid: 90b17cdf9533b1e42b2fb74bf8228c92
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
