--- Fightship Arena/Assets/Scripts/Managers/OrchestrationManagement/Spawn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9205a8b773a16394b92d02328094239f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
