--- Fightship Arena/Assets/Scripts/Managers/OrchestrationManagement/SpawnBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e6d41f2270e8fec4085eb23c71d91c76
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
