--- Fightship Arena/Assets/Scripts/Managers/OrchestrationManagement/SpawnGroup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0034b8c86035a6048abde77fc883fbdc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
