--- Fightship Arena/Assets/Scripts/Managers/OrchestrationManagement2.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0f85f3a19a223854c940f4ef6eec0d59
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
