--- Fightship Arena/Assets/Scripts/Managers/OrchestrationManagement2/OrchestrationManager2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 90b17cdf9533b1e42b2fb74bf8228c92
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
